% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_methods.R
\name{anyNA.sk}
\alias{anyNA.sk}
\title{Check for presence of grid points with missing data (NAs)}
\usage{
\method{anyNA}{sk}(x, recursive)
}
\arguments{
\item{x}{a sk object}

\item{recursive}{ignored}
}
\value{
logical
}
\description{
Returns a logical indicating if any of the grid points are NA
}
\examples{
g = sk_validate(list(gval=stats::rnorm(4^2), gdim=4, gres=0.5))
anyNA(g)
g[1] = NA
anyNA(g)

}
