% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk.R
\name{as.double.bk}
\alias{as.double.bk}
\title{Coerce grid values to numeric (double type)}
\usage{
\method{as.double}{bk}(x, ...)
}
\arguments{
\item{x}{a bk object}

\item{mode}{passed to as.double}
}
\value{
a "bk" object with numeric data values
}
\description{
This also adds support for as.numeric
}
\examples{
g = bk_validate(list(gval=sample(c(F,T), 4^2, replace=TRUE), gdim=4, gres=0.5))
g[]
as.numeric(g)[]

}
