% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_methods.R
\name{as.double.sk}
\alias{as.double.sk}
\title{Coerce grid values to numeric (double type)}
\usage{
\method{as.double}{sk}(x, ...)
}
\arguments{
\item{x}{a sk object}

\item{...}{further arguments to as.double}
}
\value{
an "sk" object with numeric data values
}
\description{
This also adds support for as.numeric
}
\examples{
g = sk_validate(list(gval=sample(c(FALSE, TRUE), 4^2, replace=TRUE), gdim=4, gres=0.5))
g[]
as.numeric(g)[]

}
