% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk.R
\name{as.integer.bk}
\alias{as.integer.bk}
\title{Coerce grid values to integer}
\usage{
\method{as.integer}{bk}(x, ...)
}
\arguments{
\item{x}{a bk object}

\item{mode}{passed to as.vector}
}
\value{
a "bk" object with integer data values
}
\description{
Coerce grid values to integer
}
\examples{
g = bk_validate(list(gval=rnorm(4^2), gdim=4, gres=0.5))
g[]
as.integer(g)[]
}
