% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_methods.R
\name{as.logical.sk}
\alias{as.logical.sk}
\title{Coerce grid values to logical}
\usage{
\method{as.logical}{sk}(x, ...)
}
\arguments{
\item{x}{a sk object}

\item{...}{further arguments to as.logical}
}
\value{
a "sk" object with logical data values
}
\description{
Coerce grid values to logical
}
\examples{
g = sk_validate(list(gval=sample(c(0,1), 4^2, replace=TRUE), gdim=4, gres=0.5))
g[]
as.logical(g)[]

# "range" for logical is reported as integer
summary(as.logical(g))

}
