% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk_export.R
\name{bk_coords}
\alias{bk_coords}
\title{Return coordinates of a grid of points in column-vectorized order}
\usage{
bk_coords(g, out = "matrix", corner = FALSE, na_omit = FALSE, quiet = FALSE)
}
\arguments{
\item{g}{any object accepted by \code{bk}}

\item{out}{character indicating return value type, either 'list', 'matrix', or 'sf'}

\item{na_omit}{logical, indicates to return only locations of observed points}

\item{corners}{logical, indicates to only return the corner points}
}
\value{
a matrix, list, or sf POINT collection, in column vectorized order
}
\description{
Expands a set of y and x grid line numbers in the column-vectorized order returned
by \code{bk}. This is similar to \code{base::expand.grid} but with the first dimension (y)
descending instead of ascending.
}
\details{
\code{out='sf'} returns an \code{sf} simple features object containing points in the same order,
with data (if any) copied from \code{g[['gval']]} into column 'gval'. Note that \code{length(g)}
points are created, which can be slow for large grids.

If \code{na_omit} is \code{TRUE} the function omits points with \code{NA} data (in \code{gval}) and only
returns the coordinates for observations. This argument is ignored when \code{corners=TRUE}
(which always returns the four corner points) or when the grid contains no observations
(all points returned).
}
\examples{
gdim = c(5,3)
g_complete = bk(gdim=gdim, gres=c(0.5, 0.7), gval=seq(prod(gdim)))
bk_coords(g_complete)
bk_coords(g_complete, out='list')

# missing data example
idx_obs =  sort(sample.int(length(g_complete), 5))
g = bk(gdim=gdim, gres=c(0.5, 0.7))
g[idx_obs] = g_complete[idx_obs]
all.equal(bk_coords(g, na_omit=TRUE), bk_coords(g_complete)[idx_obs,])

# corner points
bk_coords(g, corner=TRUE)
bk_coords(g, corner=TRUE, out='list')

# repeat with multi-layer example
g_multi = bk(modifyList(g, list(gval = cbind(g[], 2*g[]))))
all.equal(bk_coords(g_multi, na_omit=TRUE), bk_coords(g_complete)[idx_obs,])
bk_coords(g_multi, corner=TRUE)

# sf output type
if( requireNamespace('sf') ) {

# gather all points but don't copy data
sf_coords_all = bk_coords(bk(g, vals=FALSE), out='sf')

# extract non-NA data
sf_coords = bk_coords(g, out='sf', na_omit=TRUE)

# plot everything together
plot(g, reset=FALSE)
plot(sf_coords_all, add=TRUE)
plot(sf_coords, pch=16, cex=2, add=TRUE)

}

}
\seealso{
Other exporting functions: 
\code{\link{bk_export}()}
}
\concept{exporting functions}
