% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk_pars.R
\name{bk_kp}
\alias{bk_kp}
\title{Return named vector of Kronecker covariance parameters initialized to NA}
\usage{
bk_kp(k)
}
\arguments{
\item{k}{character, the kernel name, one of 'exp', 'gau', 'sph', 'gxp', 'mat'}
}
\value{
named vector of NAs, a placeholder for kernel parameters
}
\description{
Convenience function for looking up the number of parameters for a given
Kronecker covariance model, and their names. Returns a vector of \code{NA}s that
can be used as a placeholder in kernel definition lists.
}
\examples{

# there are only two possible return values
bk_kp('gau')
bk_kp('mat')
}
\seealso{
Other parameter managers: 
\code{\link{bk_bds}()},
\code{\link{bk_pars_make}()},
\code{\link{bk_pars_update}()},
\code{\link{bk_pars}()},
\code{\link{bk_to_string}()}
}
\concept{parameter managers}
\keyword{internal}
