% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk_pars.R
\name{bk_pars}
\alias{bk_pars}
\title{Initialize Kronecker covariance function parameters for a bk grid}
\usage{
bk_pars(g, pars = "gau", fill = "initial")
}
\arguments{
\item{g}{list, a bk grid list (or any other object accepted by \code{bk})}

\item{pars}{character or list defining kernels accepted by \code{bk_pars_make}}
}
\value{
a list defining the Kronecker covariance parameters
}
\description{
Returns a parameter list defining the Kronecker covariance model for \code{g}, with
default initial values assigned to parameters based on the grid dimensions and sample
variance of observed data.
}
\details{
Swap \code{fill='initial'} with 'lower' and 'upper' to get default lower and upper bounds.
}
\examples{
bk_pars(g=10)
bk_pars(c(10,15))
bk_pars(c(10,15), 'mat')
bk_pars(c(10,15), 'mat', 'upper')
}
\seealso{
bk

Other parameter managers: 
\code{\link{bk_bds}()},
\code{\link{bk_kp}()},
\code{\link{bk_pars_make}()},
\code{\link{bk_pars_update}()},
\code{\link{bk_to_string}()}
}
\concept{parameter managers}
