% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk.R
\name{bk_validate}
\alias{bk_validate}
\title{VALIDATOR}
\usage{
bk_validate(g)
}
\arguments{
\item{g}{a "bk" object or a list accepted by \code{bk_make}}
}
\value{
a validated "bk" object
}
\description{
Check compatibility of entries in a bk grid object, and fill in any missing ones
}
\details{
This constructs the object and fills missing entries. It then does some sanity checks
and computes the number of missing grid points (in list entry \code{n_missing}).

The function removes/introduces \code{idx_grid} depending on whether \code{gval} is a vector
(single-layer case) or a matrix (usually a multi-layer case). If \code{idx_grid} is missing
and \code{gval} is a matrix, it is assumed to contain all grid-points (including NAs)

The function also assigns dimension names in the order 'y', 'x' (unless otherwise
specified) for \code{gdim}, \code{gres}, and \code{gyx}.
}
\examples{

# auto-print reminds users to validate
bk_validate(list(gdim=10, gres=0.5))
bk_validate(list(gval=rnorm(10^2), gdim=10, gres=0.5))
}
