% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk.R
\name{dim.bk}
\alias{dim.bk}
\title{Grid dimensions}
\usage{
\method{dim}{bk}(x)
}
\arguments{
\item{x}{a bk object}
}
\value{
integer vector
}
\description{
Returns \code{gdim}, the number of y and x grid lines, in that order.
}
\examples{
g = bk_validate(list(gval=rnorm(4^2), gdim=4, gres=0.5))
dim(g)
}
