% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_methods.R
\name{dim.sk}
\alias{dim.sk}
\title{Grid dimensions}
\usage{
\method{dim}{sk}(x)
}
\arguments{
\item{x}{a sk object}
}
\value{
integer vector
}
\description{
Returns \code{gdim}, the number of y and x grid lines, in that order.
}
\examples{
g = sk_validate(list(gval=rnorm(4^2), gdim=4, gres=0.5))
dim(g)
}
