% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_methods.R
\name{is.na.sk}
\alias{is.na.sk}
\title{Indices of grid points with missing data (NAs)}
\usage{
\method{is.na}{sk}(x)
}
\arguments{
\item{x}{a sk object}
}
\value{
a logical vector the same length as x
}
\description{
Returns a logical vector indicating which grid points have NA values assigned
}
\examples{
g = sk_validate(list(gval=rnorm(4^2), gdim=4, gres=0.5))
g[c(1,3)] = NA
is.na(g)
}
