% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk.R
\name{length.bk}
\alias{length.bk}
\title{The number of grid-points}
\usage{
\method{length}{bk}(x)
}
\arguments{
\item{x}{a bk object}
}
\value{
integer
}
\description{
Returns the total number of points in the grid, which is the product of the
number of y and x grid lines (\code{gdim}).
}
\examples{
g = bk_validate(list(gval=rnorm(4^2), gdim=4, gres=0.5))
length(g)
}
