% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_methods.R
\name{length.sk}
\alias{length.sk}
\title{The number of grid-points}
\usage{
\method{length}{sk}(x)
}
\arguments{
\item{x}{a sk object}
}
\value{
integer
}
\description{
Returns the total number of points in the grid, which is the product of the
number of y and x grid lines (\code{gdim}).
}
\examples{
g = sk_validate(list(gval=rnorm(4^2), gdim=4, gres=0.5))
length(g)
}
