% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk.R
\name{Math.bk}
\alias{Math.bk}
\title{GROUP GENERICS}
\usage{
\method{Math}{bk}(x, ...)
}
\arguments{
\item{x}{a bk object}

\item{...}{further arguments to Math}
}
\value{
the "bk" object with data values transformed accordingly
}
\description{
Math group generics
}
\details{
All except the \code{cumsum} family are supported
}
\examples{
g = bk_validate(list(gval=rnorm(4^2), gdim=4, gres=0.5))
summary(g)
summary(abs(g))
summary(exp(g))

}
