% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_methods.R
\name{Math.sk}
\alias{Math.sk}
\title{\subsection{GROUP GENERICS}{
}}
\usage{
\method{Math}{sk}(x, ...)
}
\arguments{
\item{x}{a sk object}

\item{...}{further arguments to Math}
}
\value{
the "sk" object with data values transformed accordingly
}
\description{
Math group generics
}
\details{
All except the \code{cumsum} family are supported
}
\examples{
g = sk_validate(list(gval=rnorm(4^2), gdim=4, gres=0.5))
summary(g)
summary(abs(g))
summary(exp(g))

}
