% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk.R
\name{mean.bk}
\alias{mean.bk}
\title{Calculate the mean value in a grid}
\usage{
\method{mean}{bk}(x, ...)
}
\arguments{
\item{x}{a bk object}

\item{...}{further arguments to default mean method}
}
\value{
numeric
}
\description{
This calculates the mean over all layers (if any)
}
\examples{
g = bk_validate(list(gval=rnorm(4^2), gdim=4, gres=0.5))
mean(g)

}
