% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_methods.R
\name{mean.sk}
\alias{mean.sk}
\title{Calculate the mean value in a grid}
\usage{
\method{mean}{sk}(x, ...)
}
\arguments{
\item{x}{a sk object}

\item{...}{further arguments to default mean method}
}
\value{
numeric
}
\description{
This calculates the mean over all layers (if any)
}
\examples{
g = sk_validate(list(gval=rnorm(4^2), gdim=4, gres=0.5))
mean(g) == mean(g[])

}
