% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk.R
\name{Ops.bk}
\alias{Ops.bk}
\title{Operations group generics}
\usage{
\method{Ops}{bk}(e1, e2)
}
\arguments{
\item{e1}{a "bk" object, vector or matrix}

\item{e2}{a "bk" object, vector or matrix}
}
\value{
a "bk" object (a copy of \code{e1} or \code{e2}) with modified data values
}
\description{
Applies the operation point-wise to grid data values.
}
\details{
The function extracts the grid data \code{x[]} from all bk class arguments \code{x}, prior to
calling the default method. Before returning, the result is copied back to the grid object
of the second argument (or the first, if the second is not of class "bk").

Note that the compatibility of the two arguments is not checked beyond matching
dimension (with vectors recycled as needed). This means for example you can do
operations on two grids representing different areas, so long as they have the
same \code{gdim}.
}
\examples{
g = bk_validate(list(gval=rnorm(4^2), gdim=4, gres=0.5))

# verify a trig identity using Ops and Math
summary( cos(g)^2 + sin(g)^2 )

# create a logical grid indicating points satisfying a condition
plot(g < 0)

# test negation
all( (-g)[] == -(g[]) )

}
