% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk_pars.R
\name{bk_pars}
\alias{bk_pars}
\title{Initialize covariance kernel parameters for grid \code{g}}
\usage{
bk_pars(g, pars = "gau", fill = "initial")
}
\arguments{
\item{g}{list, a blitzKrig grid definition (or object accepted by \code{bk_grid})}

\item{pars}{character or list defining kernels (in form understood by \code{bk_pars_make})}
}
\value{
a list defining separable covariance parameters
}
\description{
Returns a kernel parameter list defining the covariance model for \code{g}, with
default initial values assigned to parameters based on the grid dimensions and sample
variance of observed data.
}
\details{
Swap \code{fill='initial'} with 'lower' and 'upper' to get default lower and upper bounds.
}
\examples{
bk_pars(g=10)
bk_pars(c(10,15))
bk_pars(c(10,15), 'mat')
bk_pars(c(10,15), 'mat', 'upper')
}
