% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk_index.R
\name{bk_sub}
\alias{bk_sub}
\title{Return a sub-grid of a blitzKrig grid object}
\usage{
bk_sub(g_obs, ij = list(), mirror = FALSE, idx = FALSE)
}
\description{
Removes grid lines as specified in \code{ij} and returns the result as a blitzKrig grid
list; If \code{ij} is empty, the function attempts to find a set of grid lines that forms
a complete regular sub-grid of \code{g_obs}.
}
\details{
\code{ij} can be a list of integer vectors, named \code{i} and \code{j}, specifying grid line numbers to
remove, in ascending order. \code{ij} can also be a nested list, with the grid lines
to remove in element \code{remove}. Alternatively, users can specify the grid lines to NOT
remove in element \code{keep}, and the function will remove the others. If both \code{keep} and
\code{remove} are specified, the function ignores \code{keep} and uses \code{remove}.

Default \code{idx=FALSE} causes the function to return the sub-grid as a blitzKrig grid object.
If \code{idx=TRUE}, the function instead returns a list containing \code{keep} and \code{remove} as
specified above. This can be passed back to bk_sub when repeatedly cropping multiple
grids in the same way.

\code{mirror=TRUE} indicates to mirror the supplied selection of grid lines in \code{ij}
about the central grid line. For example \code{ij_rem=list(i=1)} removes both the first
(left-most) column and the last (right-most).

When \code{ij} is an empty \code{list()}, the function sets \code{ij} automatically using a
greedy algorithm that iteratively checks all (four) outer grid lines and removes
the one with the highest proportion of NAs. Ties are broken in clockwise order, starting
from the left vertical. The algorithm stops when the remaining sub-grid is complete.
}
