% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk.R
\name{plot.bk}
\alias{plot.bk}
\title{Heatmap plots}
\usage{
\method{plot}{bk}(x, ...)
}
\arguments{
\item{x}{a bk object}

\item{...}{other arguments passed to bk_plot}
}
\value{
nothing
}
\description{
A wrapper for bk_plot
}
\examples{
g = bk_validate(list(gval=rnorm(4^2), gdim=4, gres=0.5))
plot(g)
}
