% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_methods.R
\name{plot.sk}
\alias{plot.sk}
\title{Heatmap plots}
\usage{
\method{plot}{sk}(x, ...)
}
\arguments{
\item{x}{a sk object}

\item{...}{other arguments passed to sk_plot}
}
\value{
nothing
}
\description{
A wrapper for sk_plot
}
\examples{
g = sk_validate(list(gval=stats::rnorm(4^2), gdim=4, gres=0.5))
plot(g)
}
\seealso{
sk_plot
}
