% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk.R
\name{print.bk}
\alias{print.bk}
\title{OTHER METHODS}
\usage{
\method{print}{bk}(x, ...)
}
\arguments{
\item{x}{a bk object}

\item{...}{ignored}
}
\value{
nothing
}
\description{
Auto-printing
}
\details{
Prints dimensions and indicates if the grid has values assigned

This prints "(not validated)" if the bk object has no \code{n_missing} entry.
This is to remind users to run \code{bk_validate}.
}
\examples{
bk_make(list(gdim=10, gres=0.5))
bk_validate(bk_make(list(gdim=10, gres=0.5)))
}
