% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_methods.R
\name{print.sk}
\alias{print.sk}
\title{\subsection{OTHER METHODS}{
}}
\usage{
\method{print}{sk}(x, ...)
}
\arguments{
\item{x}{a sk object}

\item{...}{ignored}
}
\value{
nothing
}
\description{
Auto-printing
}
\details{
Prints dimensions and indicates if the grid has values assigned

This prints "(not validated)" if the sk object has no \code{n_missing} entry.
This is to remind users to run \code{sk_validate}.
}
\examples{
sk_make(list(gdim=10, gres=0.5))
sk_validate(sk_make(list(gdim=10, gres=0.5)))
}
