% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_pars.R
\name{sk_kp}
\alias{sk_kp}
\title{Return named vector of Kronecker covariance parameters initialized to NA}
\usage{
sk_kp(k)
}
\arguments{
\item{k}{character, the kernel name, one of 'exp', 'gau', 'sph', 'gxp', 'mat'}
}
\value{
named vector of NAs, a placeholder for kernel parameters
}
\description{
Convenience function for looking up the number of parameters for a given
Kronecker covariance model, and their names. Returns a vector of \code{NA}s that
can be used as a placeholder in kernel definition lists.
}
\examples{

# there are only two possible return values
sk_kp('gau')
sk_kp('mat')
}
\seealso{
Other parameter managers: 
\code{\link{sk_bds}()},
\code{\link{sk_fit}()},
\code{\link{sk_pars_make}()},
\code{\link{sk_pars_update}()},
\code{\link{sk_pars}()},
\code{\link{sk_to_string}()}
}
\concept{parameter managers}
\keyword{internal}
