% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_estim.R
\name{sk_optim}
\alias{sk_optim}
\title{Fit covariance parameters to data by maximum (profile) likelihood using optim}
\usage{
sk_optim(
  g_obs,
  pars = "gau",
  X = 0,
  iso = FALSE,
  control = list(),
  quiet = FALSE,
  log_scale = TRUE,
  method = "L-BFGS-B",
  lower = NULL,
  initial = NULL,
  upper = NULL
)
}
\arguments{
\item{g_obs}{list of form returned by \code{sk} (with entries 'gdim', 'gres', 'gval')}

\item{pars}{list of fixed kernel parameters, with NAs indicating parameters to fit}

\item{X}{numeric, vector, matrix, or NA, the mean or its linear predictors, passed to \code{sk_LL}}

\item{iso}{logical, indicating to constrain the y and x kernel parameters to be the same}

\item{control}{list, passed to \code{stats::optim}}

\item{quiet}{logical, indicating to suppress console output}

\item{...}{named arguments to pass to \code{sk_bds}}
}
\value{
sdfsdfsd
}
\description{
documentation unfinished
}
\examples{
# set up example grid and data
g_obs = sk(10)
g_obs$gval = rnorm(10^2)
sk_optim(g_obs, quiet=TRUE)

# repeat with one or more parameters fixed
pars = sk_pars_make('gau') # NA parameters list
pars$psill = 1
sk_optim(g_obs, pars, quiet=TRUE)
pars$y$kp = 1
sk_optim(g_obs, pars, quiet=TRUE)

# iso mode constrains x parameters to equal y parameters
sk_optim(g_obs, iso=T, quiet=TRUE)
sk_optim(g_obs, pars, iso=T, quiet=TRUE)

}
