% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_plot.R
\name{sk_plot_pars}
\alias{sk_plot_pars}
\title{Plot the covariance structure of a snapKrig model}
\usage{
sk_plot_pars(pars, g, simple = FALSE, ...)
}
\arguments{
\item{pars}{list of the form returned by \code{sk_pars} with entries 'y', 'x', ('eps', 'psill')}

\item{g}{any object understood by \code{sk}}

\item{...}{additional arguments passed to \code{sk_plot}}
}
\value{
the same as \code{sk_plot}
}
\description{
Visualization of the footprint of a covariance kernel as a heatmap of size
\code{g$gdim}, where each grid cell is colored according to its covariance with
the central grid point.
}
\details{
When \code{simple=FALSE} (the default), covariance parameters are printed in the
title and axis labels with values rounded to 3 decimal places. This can be
customized by passing arguments 'main', 'ylab', 'xlab' (and any others
accepted \code{sk_plot} apart from \code{gdim}).
}
\examples{
gdim = c(100, 100)
pars = sk_pars(gdim, 'mat')
sk_plot_pars(pars, gdim)

# zoom in/out by passing a grid object with suitably modified resolution
g = sk(gdim)
sk_plot_pars(pars, g)
sk_plot_pars(pars, modifyList(g, list(gres=2*g$gres)))
sk_plot_pars(pars, modifyList(g, list(gres=0.2*g$gres)))

# change other plot style settings
sk_plot_pars(pars, modifyList(g, list(gres=0.5*g$gres)), simple=TRUE)
sk_plot_pars(pars, modifyList(g, list(gres=0.5*g$gres)), minimal=TRUE, col_invert=TRUE)

}
