% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk.R
\name{[[.bk<-}
\alias{[[.bk<-}
\title{INDEXING METHODS}
\usage{
\method{[[}{`bk<-`}(x, value)
}
\arguments{
\item{x}{a bk object}

\item{value}{the replacement object}
}
\value{
a "bk" object
}
\description{
Replace a bk list element (double-bracket assign)
}
\details{
Replaces entries in the bk list object. This does no validation (since \code{bk_validate}
uses \verb{[[<-}, so there would be an infinite recursion problem) so users are advised
to pass the results to \code{bk_validate} afterwards unless they know what they're doing.
}
\examples{
# bk list elements are interrelated - for example gres must match spacing in gyx
g = bk_validate(list(gval=rnorm(10^2), gdim=10, gres=0.5))
g[['gres']] = 2 * g[['gres']]
g[['gyx']] = lapply(g[['gyx']], function(x) 2*x)
bk_validate(g)

}
