% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_methods.R
\name{[[.sk<-}
\alias{[[.sk<-}
\title{sk_methods.R
Dean Koch, 2022
methods for sk objects (grid lists)}
\usage{
\method{[[}{`sk<-`}(x, value)
}
\arguments{
\item{x}{a sk object}

\item{value}{the replacement object}
}
\value{
a "sk" object
}
\description{
\subsection{INDEXING METHODS}{
}
}
\details{
Replace a sk list element (double-bracket assign)

Replaces entries in the sk list object. This does no validation (since \code{sk_validate}
uses \verb{[[<-}, so there would be an infinite recursion problem) so users are advised
to pass the results to \code{sk_validate} afterwards unless they know what they're doing.
}
\examples{
# sk list elements are interrelated - for example gres must match spacing in gyx
g = sk_validate(list(gval=rnorm(10^2), gdim=10, gres=0.5))
g[['gres']] = 2 * g[['gres']]
g[['gyx']] = lapply(g[['gyx']], function(x) 2*x)
sk_validate(g)

}
