% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk.R
\name{summary.bk}
\alias{summary.bk}
\title{Grid summary}
\usage{
\method{summary}{bk}(x, ...)
}
\arguments{
\item{x}{a bk object}

\item{...}{ignored}
}
\value{
nothing
}
\description{
Prints detailed information about a grid
}
\details{
All dimensional information (\code{gdim}, \code{gres}, \code{gyx}) is printed in the order y, x
}
\examples{
g = bk_validate(list(gval=rnorm(4^2), gdim=4, gres=0.5))
summary(g)
g[1] = NA
summary(g)

}
