% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_methods.R
\name{Summary.sk}
\alias{Summary.sk}
\title{Summary group generics}
\usage{
\method{Summary}{sk}(..., na.rm = FALSE)
}
\arguments{
\item{...}{further arguments to Summary}

\item{na.rm}{logical, whether to remove \code{NA}s before doing the computation}
}
\value{
a "sk" object with data values
}
\description{
Computes a summary statistic of the grid data values.
}
\examples{

# make test example and verify common summary stats
g = sk(gval=1+abs(stats::rnorm(4^2)), gdim=4, gres=0.5)
sum(g) == sum(g[])
min(g) == min(g[])
max(g) == max(g[])

# list input can include scalar, vector (but first argument must be sk class)
sum(g, -g)
sum(g, -g, 0)

# calculate product in two ways
prod(g) - exp(sum(log(g)))

# check logical conditions
any(g>0)
all(g>0)

# count the number of grid points satisfying a condition
sum(g < 2)

}
