% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_methods.R
\name{as.integer.sk}
\alias{as.integer.sk}
\title{Coerce grid values to integer}
\usage{
\method{as.integer}{sk}(x, ...)
}
\arguments{
\item{x}{a sk object}

\item{...}{further arguments to as.integer}
}
\value{
an "sk" object with integer data values
}
\description{
Coerce grid values to integer
}
\examples{
g = sk_validate(list(gval=stats::rnorm(4^2), gdim=4, gres=0.5))
g[]
as.integer(g)[]
}
