% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk_export.R
\name{bk_export}
\alias{bk_export}
\title{Convert "bk" grid to SpatRaster}
\usage{
bk_export(g, template = "terra")
}
\arguments{
\item{g}{any object accepted or returned by \code{bk}}

\item{template}{character or RasterLayer/SpatRaster to set output type

Converts a vector or matrix to a SpatRaster, or RasterLayer. Multi-layer outputs are
supported for terra but not raster.}
}
\value{
a RasterLayer or SpatRaster containing the data from \code{g} (or a sub-grid)
}
\description{
Convert "bk" grid to SpatRaster
}
\examples{

if( requireNamespace('raster') ) {

# open example file as RasterLayer
r_path = system.file('external/rlogo.grd', package='raster')
r = raster::raster(r_path, band=1)
g = bk(r)

# convert back to RasterLayer and compare
r_from_g = bk_export(g, 'raster')
print(r)
print(r_from_g)

# NOTE: layer name, band number, and various other metadata are lost
all.equal(r_from_g, r)

}

# same with terra
if( requireNamespace('terra') ) {

# convert all layers
r = terra::rast(r_path)
g = bk(r)
r_from_g = bk_export(g)

# NOTE: various metadata are lost
all.equal(r_from_g, r)

}


}
\seealso{
Other exporting functions: 
\code{\link{bk_coords}()}
}
\concept{exporting functions}
