% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk.R
\name{bk_make}
\alias{bk_make}
\title{bk.R
Dean Koch, 2022
S3 class and methods for bk objects (grid lists)
A "bk" object is just a list of 3-6 objects representing a grid and its data,
similar to the "raster" and "SpatRaster" classes in terra and raster. The S3 class
doesn't accomplish anything performance-wise. We use it to link blitzKrig grid methods
to R's many generic functions (print, plot, etc) in a sensible way, and to point R's
internal generic functions (like [, [<-, and length) to look at the contents of \code{bk$gval}}
\usage{
bk_make(g)
}
\arguments{
\item{g}{list with any of the six named arguments mentioned above (\code{gdim}, etc)}
}
\value{
a "bk" object
}
\description{
At minimum, a "bk" object contains three items:
}
\details{
\itemize{
\item gdim: vector of two positive integers, the number of grid lines (n = their product)
\item gres: vector of two positive scalars, the resolution (in distance between grid lines)
\item gyx: list of two numeric vectors (lengths matching gdim), the grid line intercepts
}

This defines the grid dimensions, their spacing (resolution), and the grid extent. All
are in the order y, x, so that gdim is consistent with \code{base::dim} if we interpret the
data grid as a matrix.
\itemize{
\item gval: numeric vector or matrix, the grid data
}

In the single layer case, data are stored as a vector (the column-vectorization), with
NAs representing missing data. In the multi-layer case, the data are stored in a matrix,
with layers in columns. For the matrix case only, we use a sparse representation and omit
the NAs. This requires an additional indexing vector:
\itemize{
\item idx_grid: length-n numeric vector mapping rows of multi-layer to grid points
}

Lastly, for geo-referenced data we store the projection info as a WKT CRS
\itemize{
\item crs: character, the CRS for the grid
}

CONSTRUCTOR

Make a bk grid object

This constructs a "bk" object from a named list containing at least the element \code{gdim}.
Users can optionally provide other list elements \code{gyx}, \code{gres}, \code{gval}, \code{crs}, and
\code{idx_grid} (which must have one non-NA index for each row in \code{gval}. See description).

Typical usage is to pass the results of \code{bk_make} to \code{bk_validate}. This fills in any
missing and checks that all entries are valid and compatible with one another.

Input classes and lengths are checked before returning. \code{gdim} and \code{gres} are length-2
vectors (with y and x elements) but they can each be specified by a single number, as
shorthand to use the same value for y and x. \code{gval} should be a matrix or vector of grid
data, and \code{crs} should be a character string (the WKT representation).
}
\examples{

# auto-print reminds users to validate
bk_make(list(gdim=10, gres=0.5))

}
