% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bk_pars.R
\name{bk_pars_make}
\alias{bk_pars_make}
\title{Build a parameter list defining the 2d spatial covariance model}
\usage{
bk_pars_make(pars = "gau")
}
\arguments{
\item{pars}{character vector of kernel name(s) or list of parameters (see DETAILS)}
}
\value{
parameter list containing sub-lists 'y', 'x', and scalars 'psill' and 'eps'
}
\description{
Constructs a nested list containing expected covariance parameters for the supplied
kernel names in \code{pars}. If \code{pars} is already such a list, the function checks that
parameter names and lengths are valid and returns a copy containing only the
expected parameters, properly named, with NAs assigned to any missing parameters.
}
\details{
\code{pars} can be a kernel name ('gau', 'mat', etc), or a vector or list of two of them,
in which case the function returns a filled-out kernel definition list indicating
expected parameters.
}
\examples{
# pass a kernel name to get 2d version with NAs for all parameters
'mat' |> bk_pars_make()
# pass a vector or list of kernel names - when unnamed, order y, x is assumed
c('gau', 'mat') |> bk_pars_make()
list('gau', 'mat') |> bk_pars_make()
list(x='gau', y='mat') |> bk_pars_make()
# when missing , x kernel definition is copied from y, and vice versa
list(k='exp', kp=c(rho=1)) |> bk_pars_make()
# when unnamed, kernel range and shape parameters are assigned expected names
list(psill=1, x=list(k='mat', kp=c(1,2))) |> bk_pars_make()
# incorrectly named parameters raise errors - example:
# list(psill=1, x=list(k='exp', kp=c(foo=1))) |> bk_pars_make()
# complete parameter definition lists are returned unchanged
k_list = list(k='exp', kp=c(rho=1))
list(psill=1, eps=0, x=k_list, y=k_list) |> bk_pars_make()
}
