% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_methods.R
\name{print.sk}
\alias{print.sk}
\title{Auto-printing}
\usage{
\method{print}{sk}(x, ...)
}
\arguments{
\item{x}{a sk object}

\item{...}{ignored}
}
\value{
nothing
}
\description{
Prints dimensions and indicates if the grid has values assigned
}
\details{
This prints "(not validated)" if the sk object has no \code{is_na} entry,
to remind users to run \code{sk_validate}.
}
\examples{
sk_make(list(gdim=10, gres=0.5))
sk_validate(sk_make(list(gdim=10, gres=0.5)))
}
