% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_methods.R
\name{summary.sk}
\alias{summary.sk}
\title{Grid summary}
\usage{
\method{summary}{sk}(x, ...)
}
\arguments{
\item{x}{a sk object}

\item{...}{ignored}
}
\value{
nothing
}
\description{
Prints detailed information about a grid
}
\details{
All dimensional information (\code{gdim}, \code{gres}, \code{gyx}) is printed in the order y, x
}
\examples{
g = sk_validate(list(gval=rnorm(4^2), gdim=4, gres=0.5))
summary(g)
g[1] = NA
summary(g)

}
